move _print to py3compat
